requirements.txt
spacy
streamlit
requests
groq
python-dotenv
openai
numpy
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-text-splitters
langchain-groq
langchain-cohere
langchain-huggingface
pypdf
sentence-transformers
faiss-cpu
fastapi
uvicorn
pinecone-client
langchain-pinecone